+++ lib/extensions.dart
@@ -1,5 +1,7 @@
 library extensions;
 
+export 'extensions/double_extension.dart';
+export 'extensions/int_extension.dart';
 export 'extensions/list_extension.dart';
 export 'extensions/string_extension.dart';
 export 'extensions/widget_extension.dart';