+++ lib/theme/text_theme.dart
@@ -1,43 +1,48 @@
 import 'package:flutter/material.dart';
 
 class TextTheme {
-  // Grey text
-  /// font size 18 - grey[300] text
-  static TextStyle grey17 = TextStyle(color: Colors.grey[300], fontSize: 17.0);
+  /// This constructor prevents instantiation and extension.
+  TextTheme._();
 
-  // White text
-  /// font size 14 - white text
-  static TextStyle white14 = TextStyle(color: Colors.white, fontSize: 14.0);
+  /// White text with font sizes between 0 and 100
+  /// Example: TextTheme.white[20]
+  static List<TextStyle> white = List<TextStyle>.generate(
+    100,
+    (size) => TextStyle(
+      color: Colors.white,
+      fontSize: size * 1.0,
+    ),
+  );
 
-  /// font size 16 - white text
-  static TextStyle white16 = TextStyle(color: Colors.white, fontSize: 16.0);
+  /// White bold text with font sizes between 0 and 100
+  /// Example: TextTheme.whiteBold[20]
+  static List<TextStyle> whiteBold = List<TextStyle>.generate(
+    100,
+    (size) => TextStyle(
+      color: Colors.white,
+      fontSize: size * 1.0,
+      fontWeight: FontWeight.bold,
+    ),
+  );
 
-  /// font size 19 - white text
-  static TextStyle white19 = TextStyle(color: Colors.white, fontSize: 19.0);
+  /// Grey[300] text with font sizes between 0 and 100
+  /// Example: TextTheme.grey[20]
+  static List<TextStyle> grey = List<TextStyle>.generate(
+    100,
+    (size) => TextStyle(
+      color: Colors.grey[300],
+      fontSize: size * 1.0,
+    ),
+  );
 
-  /// font size 20 - white text
-  static TextStyle white20 = TextStyle(color: Colors.white, fontSize: 20);
-
-  /// font size 30 - white text
-  static TextStyle white30 = TextStyle(color: Colors.white, fontSize: 30);
-
-  /// font size 40 - white text
-  static TextStyle white40 = TextStyle(color: Colors.white, fontSize: 40);
-
-  // White bold text
-  /// font size 20 - white bold text
-  static TextStyle whiteBold20 = TextStyle(
-      color: Colors.white, fontSize: 20.0, fontWeight: FontWeight.bold);
-
-  /// font size 25 - white bold text
-  static TextStyle whiteBold25 = TextStyle(
-      color: Colors.white, fontSize: 25.0, fontWeight: FontWeight.bold);
-
-  /// font size 30 - white bold text
-  static TextStyle whiteBold30 =
-      TextStyle(color: Colors.white, fontSize: 30, fontWeight: FontWeight.bold);
-
-  /// font size 40 - white bold text
-  static TextStyle whiteBold40 =
-      TextStyle(color: Colors.white, fontSize: 40, fontWeight: FontWeight.bold);
+  /// Grey[300] bold text with font sizes between 0 and 100
+  /// Example: TextTheme.greyBold[20]
+  static List<TextStyle> greyBold = List<TextStyle>.generate(
+    100,
+    (size) => TextStyle(
+      color: Colors.grey[300],
+      fontSize: size * 1.0,
+      fontWeight: FontWeight.bold,
+    ),
+  );
 }